Fix libpng and libfreetype port with wasm exceptions are enabled
Sadly, even though libpng can be compiled with setjmp/longjmp support
some libraries such as openjpeg assume that setjmp support is included.

Fixes: #19001
