+++ tools/ports/libpng.py
@@ -10,7 +10,11 @@ TAG = '1.6.37'
 HASH = '2ce2b855af307ca92a6e053f521f5d262c36eb836b4810cb53c809aa3ea2dcc08f834aee0ffd66137768a54397e28e92804534a74abb6fc9f6f3127f14c9c338'
 
 deps = ['zlib']
-variants = {'libpng-mt': {'PTHREADS': 1}}
+variants = {
+  'libpng-mt': {'PTHREADS': 1},
+  'libpng-wasm-sjlj': {'SUPPORT_LONGJMP': 'wasm'},
+  'libpng-mt-wasm-sjlj': {'PTHREADS': 1, 'SUPPORT_LONGJMP': 'wasm'},
+}
 
 
 def needed(settings):
@@ -18,7 +22,12 @@ def needed(settings):
 
 
 def get_lib_name(settings):
-  return 'libpng' + ('-mt' if settings.PTHREADS else '') + '.a'
+  suffix = ''
+  if settings.PTHREADS:
+    suffix += '-mt'
+  if settings.SUPPORT_LONGJMP == 'wasm':
+    suffix += '-wasm-sjlj'
+  return f'libpng{suffix}.a'
 
 
 def get(ports, settings, shared):
@@ -35,6 +44,8 @@ def get(ports, settings, shared):
     flags = ['-sUSE_ZLIB']
     if settings.PTHREADS:
       flags += ['-pthread']
+    if settings.SUPPORT_LONGJMP == 'wasm':
+      flags.append('-sSUPPORT_LONGJMP=wasm')
 
     ports.build_port(source_path, final, 'libpng', flags=flags, exclude_files=['pngtest'], exclude_dirs=['scripts', 'contrib'])
